requirements.txt
python-telegram-bot==21.11.1
python-dotenv==1.1.1
requests==2.32.5
patool==4.0.1
yandex-cloud-ml-sdk==0.15.0
certifi==2025.8.3
httpx==0.28.1
httpcore==1.0.9
urllib3==2.5.0
charset-normalizer==3.4.3
idna==3.10
anyio==4.10.0
